requirements.txt
asgiref==3.5.2
black==22.10.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Django==4.1.3
django-autoslug==1.9.8
django-environ==0.9.0
djangorestframework==3.14.0
idna==3.4
mypy-extensions==0.4.3
pathspec==0.10.2
Pillow==9.3.0
platformdirs==2.5.4
psycopg2-binary==2.9.5
PyJWT==2.6.0
python-decouple==3.6
pytz==2022.5
razorpay==1.3.0
requests==2.28.1
sqlparse==0.4.3
tomli==2.0.1
twilio==7.15.1
tzdata==2022.5
urllib3==1.26.12
xlwt==1.3.0
